CoinService This should separate the logic to add, remove coins. So we can reuse it in different places.
